+++ internal/repository/test_category.go
@@ -7,5 +7,5 @@ import (
 
 type TestCategory interface {
 	GetAll(ctx context.Context) ([]entity.TestCategory, error)
-	GetByID(ctx context.Context, id int) (*entity.TestCategory, error)
+	GetByID(ctx context.Context, id uint64) (*entity.TestCategory, error)
 }
